Move data generator into the tests/ dir and rename it
